Fix directory structure for workshop
