+++ src/main/resources/archetype-resources/workshops/src/main/java/com/airhacks/workshops/business/registrations/boundary/RegistrationsResource.java
@@ -7,11 +7,14 @@ import javax.ejb.Stateless;
 import javax.inject.Inject;
 import javax.json.JsonArray;
 import javax.json.JsonObject;
+import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
+import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
+import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriInfo;
 
@@ -20,6 +23,8 @@ import javax.ws.rs.core.UriInfo;
  * @author airhacks.com
  */
 @Stateless
+@Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
+@Consumes({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
 @Path("registrations")
 public class RegistrationsResource {
 
@@ -39,7 +44,7 @@ public class RegistrationsResource {
 
     @GET
     @Path("{id}")
-    public Registration find(@PathParam("id") int registrationId) {
+    public Registration find(@PathParam("id") long registrationId) {
         return registrations.find(registrationId);
     }
 
@@ -55,7 +60,7 @@ public class RegistrationsResource {
     @GET
     @Path("{id}/dummy")
     public Registration dummy(@PathParam("id") int registrationId) {
-        return new Registration(true, 1, 1);
+        return new Registration(registrationId);
     }
 
 }
