--- src/main/resources/archetype-resources/workshops/src/src/main/java/com/airhacks/workshops/business/JAXRSConfiguration.java
@@ -1,13 +1,0 @@
-package com.airhacks.workshops.business;
-
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
-
-/**
- *
- * @author airhacks.com
- */
-@ApplicationPath("resources")
-public class JAXRSConfiguration extends Application {
-
-}
