--- src/main/resources/archetype-resources/workshops/src/src/main/java/com/airhacks/workshops/business/tracing/boundary/LoggerExposer.java
@@ -1,19 +1,0 @@
-package com.airhacks.workshops.business.tracing.boundary;
-
-import java.util.logging.Logger;
-import javax.enterprise.inject.Produces;
-import javax.enterprise.inject.spi.InjectionPoint;
-
-/**
- *
- * @author airhacks.com
- */
-public class LoggerExposer {
-
-    @Produces
-    public Logger expose(InjectionPoint ip) {
-        String loggerName = ip.getMember().getDeclaringClass().getName();
-        return Logger.getLogger(loggerName);
-    }
-
-}
