--- src/main/resources/archetype-resources/workshops/src/src/test/java/com/airhacks/workshops/business/tracing/boundary/LoggerInjectionSupport.java
@@ -1,19 +1,0 @@
-package com.airhacks.workshops.business.tracing.boundary;
-
-import java.util.logging.Logger;
-import javax.inject.Inject;
-
-/**
- *
- * @author airhacks.com
- */
-public class LoggerInjectionSupport {
-
-    @Inject
-    Logger logger;
-
-    public Logger getLogger() {
-        return logger;
-    }
-
-}
