--- src/main/resources/archetype-resources/workshops/src/src/test/resources/META-INF/persistence.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<persistence version="2.0" xmlns="http://java.sun.com/xml/ns/persistence" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd">
-    <persistence-unit name="integration-test" transaction-type="RESOURCE_LOCAL">
-        <class>com.airhacks.workshops.business.registrations.entity.Registration</class>
-        <properties>
-            <property name="javax.persistence.jdbc.url" value="jdbc:derby:./workshops;create=true"/>
-            <property name="javax.persistence.jdbc.driver" value="org.apache.derby.jdbc.EmbeddedDriver"/>
-            <property name="javax.persistence.schema-generation.database.action" value="drop-and-create"/>
-        </properties>
-    </persistence-unit>
-</persistence>
